fix(TDI-37655): refactor the test
