+++ components/components-jdbc/components-jdbc-runtime/src/test/java/org/talend/components/jdbc/dataprep/TDataPrepDBInputTestIT.java
@@ -15,6 +15,7 @@ package org.talend.components.jdbc.dataprep;
 import static org.junit.Assert.*;
 
 import java.io.IOException;
+import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.List;
 import java.util.Map;
@@ -30,9 +31,11 @@ import org.junit.Test;
 import org.talend.components.api.component.ComponentDefinition;
 import org.talend.components.api.component.runtime.Reader;
 import org.talend.components.jdbc.common.DBTestUtils;
+import org.talend.components.jdbc.common.SimpleDBTable;
 import org.talend.components.jdbc.dataprep.di.TDataPrepDBInputDefinition;
 import org.talend.components.jdbc.dataprep.di.TDataPrepDBInputProperties;
 import org.talend.components.jdbc.runtime.JDBCSource;
+import org.talend.components.jdbc.runtime.JdbcRuntimeUtils;
 import org.talend.components.jdbc.runtime.setting.AllSetting;
 import org.talend.daikon.NamedThing;
 import org.talend.daikon.avro.converter.IndexedRecordConverter;
@@ -45,7 +48,9 @@ public class TDataPrepDBInputTestIT {
     public static void beforeClass() throws Exception {
         allSetting = DBTestUtils.createAllSetting();
 
-        DBTestUtils.createTable(allSetting);
+        try (Connection conn = JdbcRuntimeUtils.createConnection(allSetting)) {
+            SimpleDBTable.createTestTable(conn);
+        }
     }
 
     @AfterClass
@@ -55,7 +60,10 @@ public class TDataPrepDBInputTestIT {
 
     @Before
     public void before() throws SQLException, ClassNotFoundException {
-        DBTestUtils.truncateTableAndLoadData(allSetting);
+        try (Connection conn = JdbcRuntimeUtils.createConnection(allSetting)) {
+            DBTestUtils.truncateTable(conn);
+            SimpleDBTable.loadTestData(conn);
+        }
     }
 
     @Test
@@ -63,7 +71,7 @@ public class TDataPrepDBInputTestIT {
         TDataPrepDBInputDefinition definition = new TDataPrepDBInputDefinition();
         TDataPrepDBInputProperties properties = createCommonJDBCInputProperties(definition);
 
-        properties.main.schema.setValue(DBTestUtils.createTestSchema());
+        properties.main.schema.setValue(SimpleDBTable.createTestSchema());
         properties.sql.setValue(DBTestUtils.getSQL());
 
         JDBCSource source = DBTestUtils.createCommonJDBCSource(properties);
@@ -88,7 +96,7 @@ public class TDataPrepDBInputTestIT {
         TDataPrepDBInputDefinition definition = new TDataPrepDBInputDefinition();
         TDataPrepDBInputProperties properties = createCommonJDBCInputProperties(definition);
 
-        properties.main.schema.setValue(DBTestUtils.createTestSchema());
+        properties.main.schema.setValue(SimpleDBTable.createTestSchema());
         properties.sql.setValue(DBTestUtils.getSQL());
 
         JDBCSource source = DBTestUtils.createCommonJDBCSource(properties);
@@ -96,7 +104,7 @@ public class TDataPrepDBInputTestIT {
         Schema schema = source.getEndpointSchema(null, "TEST");
         assertEquals("TEST", schema.getName().toUpperCase());
         List<Field> columns = schema.getFields();
-        DBTestUtils.testMetadata(columns);
+        SimpleDBTable.testMetadata(columns);
     }
 
     @SuppressWarnings({ "rawtypes", "unchecked" })
@@ -107,7 +115,7 @@ public class TDataPrepDBInputTestIT {
             TDataPrepDBInputDefinition definition = new TDataPrepDBInputDefinition();
             TDataPrepDBInputProperties properties = createCommonJDBCInputProperties(definition);
 
-            properties.main.schema.setValue(DBTestUtils.createTestSchema());
+            properties.main.schema.setValue(SimpleDBTable.createTestSchema());
             properties.sql.setValue(DBTestUtils.getSQL());
 
             reader = DBTestUtils.createCommonJDBCInputReader(properties);
@@ -163,7 +171,7 @@ public class TDataPrepDBInputTestIT {
         TDataPrepDBInputDefinition definition = new TDataPrepDBInputDefinition();
         TDataPrepDBInputProperties properties = createCommonJDBCInputProperties(definition);
 
-        properties.main.schema.setValue(DBTestUtils.createTestSchema());
+        properties.main.schema.setValue(SimpleDBTable.createTestSchema());
         properties.sql.setValue(DBTestUtils.getSQL());
 
         Reader reader = DBTestUtils.createCommonJDBCInputReader(properties);
